Dodatna modularizacija i bug-fixovi
